backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
supabase==2.0.2
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP client and validation
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0

# OpenAI integration
openai==1.3.7

# Utilities
python-dotenv==1.0.0
pillow==10.1.0
python-json-logger==2.0.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
